Refactor http responses
