+++ admin/json-carves.go
@@ -66,24 +66,21 @@ func jsonCarvesHandler(w http.ResponseWriter, r *http.Request) {
 	// Extract target
 	target, ok := vars["target"]
 	if !ok {
-		incMetric(metricAdminErr)
-		log.Println("error getting target")
 		incMetric(metricJSONErr)
+		log.Println("error getting target")
 		return
 	}
 	// Verify target
 	if !CarvesTargets[target] {
-		incMetric(metricAdminErr)
-		log.Printf("invalid target %s", target)
 		incMetric(metricJSONErr)
+		log.Printf("invalid target %s", target)
 		return
 	}
 	// Retrieve carves for that target
 	qs, err := queriesmgr.GetCarves(target)
 	if err != nil {
-		incMetric(metricAdminErr)
-		log.Printf("error getting query carves %v", err)
 		incMetric(metricJSONErr)
+		log.Printf("error getting query carves %v", err)
 		return
 	}
 	// Prepare data to be returned
@@ -140,15 +137,11 @@ func jsonCarvesHandler(w http.ResponseWriter, r *http.Request) {
 	// Serialize JSON
 	returnedJSON, err := json.Marshal(returned)
 	if err != nil {
-		incMetric(metricAdminErr)
-		log.Printf("error serializing JSON %v", err)
 		incMetric(metricJSONErr)
+		log.Printf("error serializing JSON %v", err)
 		return
 	}
-	incMetric(metricAdminOK)
 	// Header to serve JSON
-	w.Header().Set(utils.ContentType, utils.JSONApplicationUTF8)
-	w.WriteHeader(http.StatusOK)
-	_, _ = w.Write(returnedJSON)
+	utils.HTTPResponse(w, utils.JSONApplicationUTF8, http.StatusOK, returnedJSON)
 	incMetric(metricJSONOK)
 }
