+++ admin/utils.go
@@ -398,23 +398,8 @@ func templateMetadata(ctx contextValue, service, version string) TemplateMetadat
 	}
 }
 
-// Helper to send HTTP response
-func apiHTTPResponse(w http.ResponseWriter, cType string, code int, data interface{}) {
-	if cType != "" {
-		w.Header().Set(utils.ContentType, cType)
-	}
-	content, err := json.Marshal(data)
-	if err != nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		log.Printf("error serializing response: %v", err)
-		content = []byte("error serializing response")
-	}
-	w.WriteHeader(code)
-	_, _ = w.Write(content)
-}
-
 // Helper to handle admin error responses
 func adminErrorResponse(w http.ResponseWriter, msg string, code int, err error) {
 	log.Printf("%s: %v", msg, err)
-	apiHTTPResponse(w, utils.JSONApplicationUTF8, code, AdminResponse{Message: msg})
+	utils.HTTPResponse(w, utils.JSONApplicationUTF8, code, AdminResponse{Message: msg})
 }
