+++ api/handlers.go
@@ -15,15 +15,6 @@ const (
 	metricHealthOK  = "health-ok"
 )
 
-// JSONApplication for Content-Type headers
-const JSONApplication string = "application/json"
-
-// ContentType for header key
-const contentType string = "Content-Type"
-
-// JSONApplicationUTF8 for Content-Type headers, UTF charset
-const JSONApplicationUTF8 string = JSONApplication + "; charset=UTF-8"
-
 var errorContent = []byte("❌")
 var okContent = []byte("✅")
 
@@ -32,7 +23,7 @@ func healthHTTPHandler(w http.ResponseWriter, r *http.Request) {
 	incMetric(metricHealthReq)
 	utils.DebugHTTPDump(r, settingsmgr.DebugHTTP(settings.ServiceAPI), true)
 	// Send response
-	apiHTTPResponse(w, JSONApplicationUTF8, http.StatusOK, okContent)
+	utils.HTTPResponse(w, "", http.StatusOK, okContent)
 	incMetric(metricHealthOK)
 }
 
@@ -41,7 +32,7 @@ func rootHTTPHandler(w http.ResponseWriter, r *http.Request) {
 	incMetric(metricHealthReq)
 	utils.DebugHTTPDump(r, settingsmgr.DebugHTTP(settings.ServiceAPI), true)
 	// Send response
-	apiHTTPResponse(w, JSONApplicationUTF8, http.StatusOK, okContent)
+	utils.HTTPResponse(w, "", http.StatusOK, okContent)
 	incMetric(metricHealthOK)
 }
 
@@ -50,7 +41,7 @@ func errorHTTPHandler(w http.ResponseWriter, r *http.Request) {
 	incMetric(metricAPIReq)
 	utils.DebugHTTPDump(r, settingsmgr.DebugHTTP(settings.ServiceAPI), true)
 	// Send response
-	apiHTTPResponse(w, JSONApplicationUTF8, http.StatusInternalServerError, errorContent)
+	utils.HTTPResponse(w, "", http.StatusInternalServerError, errorContent)
 	incMetric(metricAPIErr)
 }
 
@@ -59,6 +50,6 @@ func forbiddenHTTPHandler(w http.ResponseWriter, r *http.Request) {
 	incMetric(metricAPIReq)
 	utils.DebugHTTPDump(r, settingsmgr.DebugHTTP(settings.ServiceAdmin), true)
 	// Send response
-	apiHTTPResponse(w, JSONApplicationUTF8, http.StatusForbidden, errorContent)
+	utils.HTTPResponse(w, "", http.StatusForbidden, errorContent)
 	incMetric(metricAPIErr)
 }
